editor circle dep fix
