+++ packages/ui/components/editor/Editor.tsx
@@ -12,7 +12,6 @@ import { MarkdownShortcutPlugin } from "@lexical/react/LexicalMarkdownShortcutPl
 import { RichTextPlugin } from "@lexical/react/LexicalRichTextPlugin";
 import { HeadingNode, QuoteNode } from "@lexical/rich-text";
 import { TableCellNode, TableNode, TableRowNode } from "@lexical/table";
-import type { Dispatch, SetStateAction } from "react";
 
 import classNames from "@calcom/ui/classNames";
 
@@ -25,6 +24,7 @@ import PlainTextPlugin from "./plugins/PlainTextPlugin";
 import ToolbarPlugin from "./plugins/ToolbarPlugin";
 import CustomEnterKeyPlugin from "./plugins/customEnterKeyPlugin";
 import "./stylesEditor.css";
+import type { TextEditorProps } from "./types";
 
 /*
  Default toolbar items:
@@ -33,22 +33,6 @@ import "./stylesEditor.css";
   - italic
   - link
 */
-export type TextEditorProps = {
-  getText: () => string;
-  setText: (text: string) => void;
-  excludedToolbarItems?: string[];
-  variables?: string[];
-  addVariableButtonTop?: boolean;
-  height?: string;
-  maxHeight?: string;
-  placeholder?: string;
-  disableLists?: boolean;
-  updateTemplate?: boolean;
-  firstRender?: boolean;
-  setFirstRender?: Dispatch<SetStateAction<boolean>>;
-  editable?: boolean;
-  plainText?: boolean;
-};
 
 const editorConfig = {
   theme: ExampleTheme,
