+++ packages/ui/components/editor/plugins/ToolbarPlugin.tsx
@@ -30,7 +30,7 @@ import { createPortal } from "react-dom";
 import { Button } from "../../button";
 import { Dropdown, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "../../dropdown";
 import { Icon } from "../../icon";
-import type { TextEditorProps } from "../Editor";
+import type { TextEditorProps } from "../types";
 import { AddVariablesDropdown } from "./AddVariablesDropdown";
 
 const LowPriority = 1;
